ci: validate commit messages
I added NPM dependencies on commitlint and Husky. commitlint is being
used to enforce that commit messages are properly formatted and follow
the Conventional Commits specification. Husky is being used to run
commitlint automatically to validate commit messages before they are
persisted to the repository.

I created a GitHub Actions workflow that will use commitlint to validate
commit messages. The workflow will validate all commit messages that are
pushed for a pull request. On the main branch, the workflow will
validate the commit message for the commit that was merged into the
main branch.

I added commitizen as an NPM dependency. I configured the
prepare-commit-msg Git hook to use commitizen to help the developer to
write a properly formatted commit message.