build(ios): create iOS application
I created the Naked Standup application for iOS and iPadOS. I created
the Tuist manifest files that are used to generate the Xcode workspace,
project, and the application target. I created a basic SwiftUI
application to make sure that the application ran in the simulator and
on a device.

I added a dependency on Fastlane Tools. Fastlane Tools is used to manage
the code signing certificates and provisioning profiles for Apple, to
run the mobile unit tests, and to build and publish the application to
the Apple App Store.

I implemented the GitHub Actions workflow and Fastlane lanes for
building and publishing Naked Standup to Apple's TestFlight service.
TestFlight can be used by early adopters to test Naked Standup and
provide feedback for release.